SCHED-304: Separated logic out for ResourceManager and fixed configuration management.
